added fortran module and a suggestion for how to test
